login logout

+++ src/main/java/vn/hoidanit/jobhunter/controller/UserController.java
@@ -12,12 +12,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.turkraft.springfilter.boot.Filter;
 
 import vn.hoidanit.jobhunter.domain.User;
+import vn.hoidanit.jobhunter.domain.dto.CreateUserDTO;
+import vn.hoidanit.jobhunter.domain.dto.FetchUserDTO;
 import vn.hoidanit.jobhunter.domain.dto.ResultPaginationDTO;
+import vn.hoidanit.jobhunter.domain.dto.UpdateUserDTO;
 import vn.hoidanit.jobhunter.service.UserService;
 import vn.hoidanit.jobhunter.util.annotation.ApiMessage;
 import vn.hoidanit.jobhunter.util.error.IdInvalidException;
@@ -25,6 +29,7 @@ import vn.hoidanit.jobhunter.util.error.IdInvalidException;
 import org.springframework.web.bind.annotation.PutMapping;
 
 @RestController
+@RequestMapping("/api/v1")
 public class UserController {
     private final UserService userService;
     private final PasswordEncoder passwordEncoder;
@@ -36,28 +41,57 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    public ResponseEntity<User> createNewUser(@RequestBody User user) {
+    @ApiMessage("Create a User")
+    public ResponseEntity<CreateUserDTO> createNewUser(@RequestBody User user) throws IdInvalidException {
+        if (this.userService.CheckExistEmail(user.getEmail())) {
+            throw new IdInvalidException("Email đã tồn tại");
+        }
         String hashPassword = passwordEncoder.encode(user.getPassword());
         user.setPassword(hashPassword);
         User crrUser = this.userService.handleSaveUser(user);
-        return ResponseEntity.status(HttpStatus.CREATED).body(crrUser);
+        CreateUserDTO userDTO = new CreateUserDTO();
+        userDTO.setId(crrUser.getId());
+        userDTO.setAddress(crrUser.getAddress());
+        userDTO.setAge(crrUser.getAge());
+        userDTO.setEmail(crrUser.getEmail());
+        userDTO.setGender(crrUser.getGender());
+        userDTO.setName(crrUser.getName());
+        userDTO.setCreatedAt(crrUser.getCreatedAt());
+        return ResponseEntity.status(HttpStatus.CREATED).body(userDTO);
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable("id") long id)
+    @ApiMessage("Delete a User")
+    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id)
             throws IdInvalidException {
+        if (!this.userService.CheckExistId(id)) {
+            throw new IdInvalidException("User có id = " + id + " không tồn tại");
+        }
         if (id >= 1500) {
             throw new IdInvalidException("id khong lon hon 1500");
         }
         this.userService.handleDeleteUser(id);
-        return ResponseEntity.ok("success");
+        return ResponseEntity.status(HttpStatus.OK).body(null);
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable("id") long id) {
+    @ApiMessage("Fetch a User")
+    public ResponseEntity<FetchUserDTO> getUserById(@PathVariable("id") long id) throws IdInvalidException {
+        if (!this.userService.CheckExistId(id)) {
+            throw new IdInvalidException("User có id = " + id + " không tồn tại");
+        }
         Optional<User> user = this.userService.fetchUserById(id);
         if (user.isPresent()) {
-            return ResponseEntity.ok(user.get());
+            FetchUserDTO userDTO = new FetchUserDTO();
+            userDTO.setId(user.get().getId());
+            userDTO.setAddress(user.get().getAddress());
+            userDTO.setAge(user.get().getAge());
+            userDTO.setEmail(user.get().getEmail());
+            userDTO.setGender(user.get().getGender());
+            userDTO.setName(user.get().getName());
+            userDTO.setCreatedAt(user.get().getCreatedAt());
+            userDTO.setUpdatedAt(user.get().getUpdatedAt());
+            return ResponseEntity.ok(userDTO);
         }
         return null;
     }
@@ -70,8 +104,11 @@ public class UserController {
     }
 
     @PutMapping("/users")
-    public ResponseEntity<User> putUser(@RequestBody User user) {
-
+    @ApiMessage("Update a User")
+    public ResponseEntity<UpdateUserDTO> putUser(@RequestBody User user) throws IdInvalidException {
+        if (!this.userService.CheckExistId(user.getId())) {
+            throw new IdInvalidException("User có id = " + user.getId() + " không tồn tại");
+        }
         return ResponseEntity.ok(this.userService.handleUpdateUser(user));
     }
